Update vdir.rb
Line 39:     add error code 183 for appcmd.exe to ignore errors when the vdir already exists.

When you create a vdir under an application <site>\<application>\<vdir> and run chef-client twice, it will return the 183 error causing the chef-run to fail. This does not occur when the vdir is only one level down, as in <site>\<vdir>.

The iis_vdir provider does not take a "returns [x,y,z]" attribute so the error code must be added to the vdir.rb provider.