clear app pool schedule in the case that there is more than one schedule
Because this functionality can add one and only one app pool recycle
schedule, we want to make sure it stays that way.  If there is more than
one schedule, clear out the existing schedules and set
is_new_recycle_at_time to true so the schedule can be reset to what we
want.